engine/requirements.txt
grpcio>=1.60.0
grpcio-tools>=1.60.0
protobuf>=4.25.0
numpy>=1.24.0
librosa>=0.10.0
